d29_frontend/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
google-api-core==2.0.1
google-auth==2.1.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.1
google-crc32c
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
munch==2.5.0
packaging==21.0
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
requests==2.26.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
